Caves and Ravines: fixed compilation and linking issues on gcc
git-svn-id: http://mc-server.googlecode.com/svn/trunk@699 0a769ca7-a7f5-676a-18bf-c427514a06d6

+++ source/Ravines.cpp
@@ -450,7 +450,7 @@ void cStructGenRavines::cRavine::ProcessChunk(
 			int DistSq = (DifX + x) * (DifX + x) + (DifZ + z) * (DifZ + z);
 			if (DistSq <= RadiusSq)
 			{
-				int Top = std::min(itr->m_Top, cChunkDef::Height);
+				int Top = std::min(itr->m_Top, (int)(cChunkDef::Height));  // Stupid gcc needs int cast
 				for (int y = std::max(itr->m_Bottom, 1); y <= Top; y++)
 				{
 					cChunkDef::SetBlock(a_BlockTypes, x, y, z, E_BLOCK_AIR);
